Split icon metadata into files
