+++ src/meta/icons/index.ts
@@ -1,0 +1,5 @@
+export interface IconMeta {
+  color?: string
+  fill?: boolean
+  svg: any
+}
