+++ src/meta/icons/seek-logo.ts
@@ -1,0 +1,75 @@
+import { IconMeta } from "."
+
+const GitLabSVG = require("../../assets/seek-logo/gitlab.svg")
+const PayPalSVG = require("../../assets/seek-logo/paypal.svg")
+const AngularSVG = require("../../assets/seek-logo/angular.svg")
+const BootstrapSVG = require("../../assets/seek-logo/bootstrap.svg")
+const DjangoSVG = require("../../assets/seek-logo/django.svg")
+const GitSVG = require("../../assets/seek-logo/git.svg")
+const GatsbySVG = require("../../assets/seek-logo/gatsby.svg")
+const FirebaseSVG = require("../../assets/seek-logo/firebase.svg")
+const HTML5SVG = require("../../assets/seek-logo/html5.svg")
+const NetlifySVG = require("../../assets/seek-logo/netlify.svg")
+const PythonSVG = require("../../assets/seek-logo/python.svg")
+const TypeScriptSVG = require("../../assets/seek-logo/typescript.svg")
+const VSCodeSVG = require("../../assets/seek-logo/vscode.svg")
+
+export const angularIconMeta: IconMeta = {
+  color: "#D6002F",
+  svg: AngularSVG,
+}
+
+export const bootstrapIconMeta: IconMeta = {
+  color: "#533B78",
+  svg: BootstrapSVG,
+}
+
+export const djangoIconMeta: IconMeta = {
+  svg: DjangoSVG,
+}
+
+export const firebaseIconMeta: IconMeta = {
+  svg: FirebaseSVG,
+}
+
+export const gatsbyIconMeta: IconMeta = {
+  color: "#E94E31",
+  svg: GatsbySVG,
+}
+
+export const gitIconMeta: IconMeta = {
+  color: "#E94E31",
+  svg: GitSVG,
+}
+
+export const gitLabIconMeta: IconMeta = {
+  color: "#F46A25",
+  svg: GitLabSVG,
+}
+
+export const html5IconMeta: IconMeta = {
+  color: "#F64A1D",
+  svg: HTML5SVG,
+}
+
+export const netlifyIconMeta: IconMeta = {
+  svg: NetlifySVG,
+}
+
+export const payPalIconMeta: IconMeta = {
+  color: "#002E82",
+  svg: PayPalSVG,
+}
+
+export const pythonIconMeta: IconMeta = {
+  svg: PythonSVG,
+}
+
+export const typeScriptIconMeta: IconMeta = {
+  color: "#356A97",
+  svg: TypeScriptSVG,
+}
+
+export const vsCodeIconMeta: IconMeta = {
+  svg: VSCodeSVG,
+}
